# packages/logstack-go/openapi.yaml
openapi: 3.1.0
info:
  title: Logstack API
  description: |
    Production-ready log management platform API.

    ## Authentication

    Logstack uses two authentication methods:

    - **JWT Bearer Tokens**: For user-facing operations (projects, alerts, settings)
    - **API Keys**: For log ingestion from your applications

    ## Rate Limits

    | Endpoint Type | Limit | Window |
    |--------------|-------|--------|
    | Log Ingestion | 1000 | 1 minute |
    | API Queries | 100 | 1 minute |
    | Authentication | 10 | 1 minute |
  version: 1.0.0
  contact:
    name: Logstack Support
    url: https://logstack.tech/docs
    email: support@logstack.tech
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.logstack.tech/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Development
...
security:
  - BearerAuth: []
paths:
  /auth/signup:
    post:
      tags: [Authentication]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  minLength: 8
                  example: securepassword123
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          description: Email already registered
  /auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: Tokens refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "401":
          description: Invalid refresh token
...
  /logs:
    get:
      tags: [Logs]
      summary: Query logs
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
            enum: [info, warn, error, critical]
        - name: source
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
          description: Full-text search in message
        - name: startTime
          in: query
          schema:
            type: string
            format: date-time
        - name: endTime
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        "200":
          description: Logs retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Log"
                  meta:
                    $ref: "#/components/schemas/Pagination"
  /logs/{id}:
    get:
      tags: [Logs]
      summary: Get log by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Log found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Log"
        "404":
          description: Log not found
  /projects:
    get:
      tags: [Projects]
      summary: List projects
      responses:
        "200":
          description: Projects retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"

    post:
      tags: [Projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      tags: [Projects]
      summary: Get project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

    put:
      tags: [Projects]
      summary: Update project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

    delete:
      tags: [Projects]
      summary: Delete project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Project deleted
...
  /alerts:
    get:
      tags: [Alerts]
      summary: List alert rules
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Alerts retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Alert"

    post:
      tags: [Alerts]
      summary: Create alert rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertCreate"
      responses:
        "201":
          description: Alert created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{id}:
    put:
      tags: [Alerts]
      summary: Update alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertUpdate"
      responses:
        "200":
          description: Alert updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"

    delete:
      tags: [Alerts]
      summary: Delete alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Alert deleted
  /alerts/options:
    get:
      tags: [Alerts]
      summary: Get alert configuration options
      description: Returns available channels, trigger patterns, trigger levels, and cooldown options for building alert rule forms.
      responses:
        "200":
          description: Alert options retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertOptions"
        "401":
          description: Unauthorized
...
  /auth/qr/generate:
    post:
      tags: [Authentication]
      summary: Generate a QR code for login
      description: Generates a QR code image and session token. The web user must be authenticated. The mobile user scans the QR and confirms via `POST /auth/qr/{token}/confirm`.
      responses:
        "200":
          description: QR session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QRGenerateResponse"
        "401":
          description: Unauthorized
  /auth/qr/{token}/status:
    get:
      tags: [Authentication]
      summary: Poll QR session status
      description: Returns the current status of a QR login session (`pending` or `confirmed`). Returns 410 if the session has expired.
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: QR session token
      responses:
        "200":
          description: Session status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [pending, confirmed]
        "410":
          description: QR session expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: QR_EXPIRED
                message: QR session has expired
  /auth/qr/{token}/confirm:
    post:
      tags: [Authentication]
      summary: Confirm a QR login session
      description: Called by the mobile app after scanning the QR code. Validates credentials and marks the QR session as confirmed, returning a JWT token pair to the mobile caller.
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: QR session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: QR session confirmed — JWT tokens returned to mobile caller
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "401":
          description: Invalid credentials
        "409":
          description: QR session already used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: QR_ALREADY_USED
                message: This QR code has already been used
        "410":
          description: QR session expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: QR_EXPIRED
                message: QR session has expired
  /auth/accept-invite:
    get:
      tags: [Authentication]
      summary: Accept an organization invite
      description: Looks up the invite by token, validates expiry, creates an OrganizationMember record, and returns JWT tokens. Creates the user account if they don't exist yet.
      security: []
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: Invite token from the invitation email
      responses:
        "200":
          description: Invite accepted — JWT tokens returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "410":
          description: Invite expired or not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: INVITE_EXPIRED
                message: This invite has expired
  /organizations/{id}/invites:
    post:
      tags: [Organizations]
      summary: Create an organization invite
      description: Sends an invite email to the specified address. Caller must be org owner or admin. Requires `team_management` feature (pro+ tier).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Organization ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, role]
              properties:
                email:
                  type: string
                  format: email
                role:
                  type: string
                  enum: [member, admin]
      responses:
        "201":
          description: Invite created and email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invite"
        "402":
          description: Subscription upgrade required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: UPGRADE_REQUIRED
                message: This feature requires a higher subscription tier.
                upgradeUrl: /checkout
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags: [Organizations]
      summary: List pending invites for an organization
      description: Returns all invites (pending, accepted, revoked) for the org. Caller must be owner or admin.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Organization ID
      responses:
        "200":
          description: Invites retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  invites:
                    type: array
                    items:
                      $ref: "#/components/schemas/Invite"
        "402":
          description: Subscription upgrade required
        "403":
          description: Insufficient role
  /organizations/{id}/invites/{inviteId}:
    delete:
      tags: [Organizations]
      summary: Revoke an organization invite
      description: Deletes a pending invite. Caller must be owner or admin. Returns 404 if the invite doesn't exist.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Organization ID
        - name: inviteId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Invite ID
      responses:
        "204":
          description: Invite revoked
        "403":
          description: Insufficient role
        "404":
          description: Invite not found
  /billing/invoices:
    get:
      tags: [Billing]
      summary: List invoices for the authenticated user
      description: Returns a paginated list of invoices ordered by `created_at DESC`.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number (20 invoices per page)
      responses:
        "200":
          description: Invoices retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoices:
                    type: array
                    items:
                      $ref: "#/components/schemas/Invoice"
                  total:
                    type: integer
                    description: Total number of invoices
                  page:
                    type: integer
        "401":
          description: Unauthorized
  /billing/invoices/{id}:
    get:
      tags: [Billing]
      summary: Get a single invoice
      description: Returns a single invoice by ID. Returns 403 if the invoice belongs to a different user.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Invoice ID
      responses:
        "200":
          description: Invoice retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "401":
          description: Unauthorized
        "403":
          description: Invoice belongs to a different user
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Project API key (starts with ls_live_ or ls_test_)
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: usr_abc123
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        accessToken:
          type: string
        refreshToken:
          type: string
    TokenPair:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
...
    Log:
      allOf:
        - $ref: "#/components/schemas/LogEntry"
        - type: object
          properties:
            id:
              type: string
            projectId:
              type: string
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        apiKey:
          type: string
        logCount:
          type: integer
        archivedAt:
          type: string
          format: date-time
          nullable: true
          description: Set when the project is archived. Null for active projects.
          example: null
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Alert:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        name:
          type: string
        condition:
          $ref: "#/components/schemas/AlertCondition"
        channels:
          type: array
          items:
            type: string
            enum: [email, push]
        cooldown:
          type: string
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    AlertCreate:
      type: object
      required: [projectId, name, condition, channels]
      properties:
        projectId:
          type: string
        name:
          type: string
        condition:
          $ref: "#/components/schemas/AlertCondition"
        channels:
          type: array
          items:
            type: string
            enum: [email, push]
        cooldown:
          type: string
          default: 15m
    AlertUpdate:
      type: object
      properties:
        name:
          type: string
        condition:
          $ref: "#/components/schemas/AlertCondition"
        channels:
          type: array
          items:
            type: string
        cooldown:
          type: string
        enabled:
          type: boolean
    AlertOptions:
      type: object
      description: Available configuration options for building alert rule forms
      properties:
        channels:
          type: array
          items:
            type: string
          example: [email, push, webhook]
        triggerPatterns:
          type: array
          items:
            type: string
          example: [".*error.*", ".*exception.*", ".*fatal.*", ".*critical.*", ".*timeout.*", ".*panic.*"]
        triggerLevels:
          type: array
          items:
            type: string
          example: [debug, info, warn, error, critical, fatal]
        cooldownOptions:
          type: array
          items:
            type: integer
          example: [5, 10, 15, 30, 60]
...
    QRGenerateResponse:
      type: object
      description: Response from generating a QR login session
      properties:
        token:
          type: string
          format: uuid
          description: The QR session token (also encoded in the QR image)
          example: "550e8400-e29b-41d4-a716-446655440000"
        qrImageUrl:
          type: string
          description: Base64-encoded PNG data URL of the QR code image
          example: "data:image/png;base64,iVBORw0KGgo..."
    ErrorResponse:
      type: object
      description: Structured error response returned by new endpoints
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: QR_EXPIRED
        message:
          type: string
          description: Human-readable error description
          example: QR session has expired
        upgradeUrl:
          type: string
          description: Present only for UPGRADE_REQUIRED errors
          example: /checkout
...
    Invite:
      type: object
      description: An organization membership invite
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440002"
        organizationId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440003"
        email:
          type: string
          format: email
          example: newmember@example.com
        role:
          type: string
          enum: [member, admin]
          example: member
        status:
          type: string
          enum: [pending, accepted, revoked]
          example: pending
        expiresAt:
          type: string
          format: date-time
          description: Invite expires 48 hours after creation
          example: "2024-01-17T09:00:00Z"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T09:00:00Z"
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: VALIDATION_ERROR
              message: Invalid request body
              details:
                - field: email
                  message: Invalid email format
...